+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/InstallerModule.java
@@ -8,11 +8,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.vyarus.dropwizard.guice.module.context.ConfigurationContext;
 import ru.vyarus.dropwizard.guice.module.context.info.impl.ExtensionItemInfoImpl;
+import ru.vyarus.dropwizard.guice.module.context.option.Options;
 import ru.vyarus.dropwizard.guice.module.context.stat.Stat;
 import ru.vyarus.dropwizard.guice.module.installer.install.binding.BindingInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.install.binding.LazyBinding;
 import ru.vyarus.dropwizard.guice.module.installer.internal.ExtensionsHolder;
 import ru.vyarus.dropwizard.guice.module.installer.internal.FeatureInstallerExecutor;
+import ru.vyarus.dropwizard.guice.module.installer.option.WithOptions;
 import ru.vyarus.dropwizard.guice.module.installer.order.OrderComparator;
 import ru.vyarus.dropwizard.guice.module.installer.scanner.ClassVisitor;
 import ru.vyarus.dropwizard.guice.module.installer.scanner.ClasspathScanner;
@@ -27,10 +29,15 @@ import static ru.vyarus.dropwizard.guice.module.context.stat.Stat.InstallersTime
 /**
  * Module performs auto configuration using classpath scanning or manually predefined installers and beans.
  * First search provided packages for registered installers
- * {@link FeatureInstaller}. Then scan classpath obe more time with installers to apply extensions.
- * <p>Feature installers can be disabled from bundle config.</p>
+ * {@link FeatureInstaller}. Then scan classpath one more time with installers to apply extensions.
+ * <p>
+ * Feature installers can be disabled from bundle config.
+ * <p>
  * NOTE: Classpath scan will load all found classes in configured packages. Try to reduce scan scope
  * as much as possible.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 31.08.2014
  */
 public class InstallerModule extends AbstractModule {
     private static final OrderComparator COMPARATOR = new OrderComparator();
@@ -98,10 +105,15 @@ public class InstallerModule extends AbstractModule {
     private List<FeatureInstaller> prepareInstallers(
             final List<Class<? extends FeatureInstaller>> installerClasses) {
         final List<FeatureInstaller> installers = Lists.newArrayList();
+        // different instance then used in guice context, but it's just an accessor object
+        final Options options = new Options(context.options());
         for (Class<? extends FeatureInstaller> installerClass : installerClasses) {
             try {
                 final FeatureInstaller installer = installerClass.newInstance();
                 installers.add(installer);
+                if (WithOptions.class.isAssignableFrom(installerClass)) {
+                    ((WithOptions) installer).setOptions(options);
+                }
                 if (logger.isTraceEnabled()) {
                     logger.trace("Registered installer: {}", FeatureUtils.getInstallerExtName(installerClass));
                 }